refactor!: make #to_s header string less concrete
Before it relied on the board having 7 columns, now it calculates the
number of columns and produces the correct result. I will not be adding
a test for different column numbers though, as I am only interested in
testing current functionality. But the code is more open to change.

Also, changed the grid loop from #each to #map which I prefer for
producing arrays from other arrays.

Looking back, I think in practice this was not worth the time actually.
The code is arguably better like this though, so I shall leave it as is.

BREAKING CHANGE: Any grid of different dimensions to the default may
have different results from #to_s now.